src/data_enricher/v1/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
chompjs==1.2.4
dateparser==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
outcome==1.3.0.post0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
selenium==4.23.1
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.32
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
ua-generator==0.5.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
wrapworks==1.0.3
wsproto==1.2.0
